apiext: wait for CRD using v1beta1 for compatibility on upgrades
This is softening 4ee7b07 (#259) to use the older v1beta1 API version
for waiting for CRDs to be established. This fixes the creation of
v1beta1 CRDs before kube-apiserver is updated during a 4.2->4.3
update.

Note: this will not fix the creation of v1 CRDs before kube-apiserver is
upgraded. Hence, we are still restricted in the use of v1 CRDs
for early manifests.