pkg/cvo: Reason granularity for RemoteFailed
We want to be able to distinguish these conditions, which can be due
to internal misconfiguration or external Cincinnati/network errors
[1].  The former can be fixed by cluster admins.  The latter could go
either way.

I dropped the len(upstream) guard from checkForUpdate because there's
already an earlier guard in syncAvailableUpdates.  The guard I'm
removing is from db150e6db4 (cvo: Perform status updates in a single
thread, 2018-11-03, #45).  The covering guard is from the later
286641de68 (api: Update to objects from openshift/api, 2018-11-15, #55).

Personally, I'd rather have GetUpdates return an *Error, so we could
dispense with the cast and unused Unknown-reason fallback.  But
Abhinav wanted the explicit cast in return for a more familiar error
type [2].

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1685338
[2]: https://github.com/openshift/cluster-version-operator/pull/268#discussion_r344823027