Add proxy support to CVO image verification.
This commit adds support for using the cluster proxy and trust bundle to verify image signatures in the CVO. Before this commit, the CVO would create the release verifier and store inside the verifier an http client for all traffic. The verifier is created before the proxy lister and values have been established by the operator, so those values cannot be read in order to properly configure the client; furthermore, the client needs to take into account that the cluster proxy is subject to change. To address these problems, the verifier has been updated to accept an interface which can provide an up-to-date client whenever a client is needed. The initial creation of the verifier accepts a reference to the operator to implement the interface.