pkg/cvo: Fetch proxy CA certs from openshift-config-managed/trusted-ca-bundle
The API docs [1] recommend avoiding trustedCA (unless you happen to be
the "proxy validator") and instead pulling the trust bundle from this
managed namespace.  The docs also explain that the trusted-ca-bundle
ConfigMap has already been merged with the system certificates, so we
don't need to inject those locally.  If trusted-ca-bundle doesn't
exist, we'll fall back to our local system store.

Our logic was touched most recently in 5968cdf933 (pkg: switch to
openshift-config for proxy CA, 2019-08-07, #231), which resolved the
"looking in the wrong place" issue by looking at the trustedCA source
(which feeds the proxy validator).  With this commit we switch that
around and look at the proxy validator's output.

[1]: https://github.com/openshift/api/blob/f2a771e1a90ceb4e65f1ca2c8b11fc1ac6a66da8/config/v1/types_proxy.go#L44-L52