pkg/start: Drop the internal EnableMetrics
We've had it since 2b81f47626 (cvo: Release our leader lease when we
are gracefully terminated, 2019-01-16, #87), but it's redundant
vs. "ListenAddr is not an empty string".

I'm also switching to:

  o.ListenAddr != ""

instead of:

  len(o.ListenAddr) > 0

because it seems slightly easier to understand, but obviously either
will work.

Cherry-picked from 07e5809d54 (#349), around conflicts due to the lack
of TLS metrics in 4.5.