pkg/cvo/status: Always set status.desired to match current CVO
As described in [1], bumping status.desired when a new
spec.desiredUpdate is set leads to trouble like:

  $ oc adm upgrade --to 4.3.13
  Updating to 4.3.13
  $ oc get -o json clusterversion version | jq -r '.status.conditions[] | .lastTransitionTime + " " + .type + " " + .status + " " + .message' | sort
  2020-04-20T20:57:12Z RetrievedUpdates True
  2020-04-20T21:23:40Z Available True Done applying 4.3.10
  2020-04-20T22:01:55Z Upgradeable False Cluster operator kube-apiserver cannot be upgraded: DefaultSecurityContextConstraintsUpgradeable: Default SecurityContextConstraints object(s) have mutated [privileged]
  2020-04-20T22:16:40Z Failing True Precondition "ClusterVersionUpgradeable" failed because of "DefaultSecurityContextConstraints_Mutated": Cluster operator kube-apiserver cannot be upgraded: DefaultSecurityContextConstraintsUpgradeable: Default SecurityContextConstraints object(s) have mutated [privileged]
  2020-04-20T22:16:40Z Progressing True Unable to apply 4.3.13: it may not be safe to apply this update
  $ oc get -o json clusterversion version | jq -r '.status.desired.version'
  4.3.13
  $ oc adm upgrade --to=4.3.13 --force
  info: Cluster is already at version 4.3.13
  $ oc adm upgrade --clear
  Cleared the update field, still at 4.3.13
  $ oc get -o json clusterversion version | jq -r '.status.desired.version'
  4.3.10

Where the "already at..." and "still at ..." messages are relying on
the status.desired semantics of [2]:

  > desired is the version that the cluster is reconciling towards.

Ideally the property would have been called 'current' or some such,
but it's too late to change it now.

When the user sets spec.desiredUpdate, they are making their intention
clear (and bumping history at this point is appropriate, because we
need *somewhere* to store the 'verified' history entry).  But to match
the "reconciling towards" semantics, this commit shifts the actual
status.desired bump so it happens right after the new CVO comes up
(with the new currentVersion).

This change reverses the earlier:

  Prefers the payload version over the operator's version...

comments from 961873d66a (sync: Do config syncing in the background,
2019-01-11, #82).  It also sets the stage for a world in which [3] has
been fixed and the CVO continues to apply the current release's
manifests while vetting a new target's preconditions in parallel.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1826115
[2]: https://github.com/openshift/api/blob/0f159fee64dbf711d40dac3fa2ec8b563a2aaca8/config/v1/types_cluster_version.go#L82-L87
[3]: https://bugzilla.redhat.com/show_bug.cgi?id=1822752