Bug 1834568: Add cert to service to support https
Fix for https://bugzilla.redhat.com/show_bug.cgi?id=1809195 is to add
https metrics target. The PR to fix that bug is failing CI upgrade
test since the 0001_00_cluster-version-operator_03_service with the
service.beta.openshift.io/serving-cert-secret-name annotation is
coming too late. So the monitoring tools aren't creating the secret
causing the CVO container to fail to start.

This PR adds the annotation to the service so that change can be
fielded ahead of remaining changes necessary to support https
target.