*: Unify terminology around "upgrades" instead of "updates"
We were not very consistent before, which can be confusing for folks
who wonder if we are attempting to make a distinction between the two
terms.  While some folks try to distinguish between bugfix updates and
new-feature upgrades, the CVO does not make that distinction.  This
commit moves us to the "upgrades" where we can do so without breaking
APIs.  For example, it does not change:

* The cluster_version_available_updates metric.
* The RetrievedUpdates condition type and its reasons like
  UpdatePayloadIntegrity.
* External API's like ClusterVersion's desiredUpdate,
  availableUpdates, and updateHistory.

To avoid slippery slopes, all internal handlers for those external
APIs are being changed to use "upgrades", to minimize confusion about
where boundaries should be drawn between the external API and internal
naming.

I am not changing update-codegen.sh or update-vendor.sh, since those
are internal tools that are distinct from cluster upgrades.

I am changing --enable-auto-update -> --enable-auto-upgrade, because
while that is technically an external API, the only consumers who
could possibly touch it today are folks who are overriding the CVO
deployment, and those folks are unlikely to be upgrading anyway.