doc/dev/upgrades: Add a blurb about restarting upgrades
I was going to make this a blog entry, but it got shot down
as sounding like a bug and we didn't want customers to worry.

Let's stash it here in upstream git where few customers
will find it, and the ones that do are hopefully more technically
inclined and will understand the logic.

(Originally stored at https://hackmd.io/Yph-TnRmR9ekzEJht0IzUA )