verify: Support verification of signatures in container/image format
Containers image expects signature stores to have the structure

    <root>/signatures/<repository>@<sha_with_=_separator>/signature-<N>

for example:

    https://mirror.openshift.com/pub/openshift-v4/signatures/openshift-release-dev/ocp-release-nightly@sha256=0a1e7de904f7b2b8d635a5018c4f15ea68cd4bbae3908cb7e788b9f2c8e66e1f/signature-1

Allow the verifier to address repositories in this format by checking the
current structure `<repo>/<digest>/signature-<N>` and then checking a new form
`<repo>@<digest>/signature-<N>`. This will allow us to publish stores in the
same form as the container/image spec in the future, while keeping backwards
compatibility with the existing structure.