lib/resourcebuilder/apps: Explicit mode instead of 'Generation > 1'
deploymentBuilder and daemonsetBuilder grow mode properties.  They had
been using 'actual.Generation > 1' as a proxy for "post-install" since
14fab0b2c8 (add generic 2-way merge handler for random types,
2018-09-27, #26), but generation 1 is just "we haven't changed the
object since it was created", not "we're installing a fresh cluster".
For example, a new Deployment or DaemonSet could be added as part of a
cluster update, and we don't want special install-time "we don't care
about specific manifest failures" then.