Bug 1822513: Determine current version by checking for status completed
Preconditions such as ClusterVersion overrides should block all upgrades
including z-level upgrades. However other preconditions should not be
block z-level upgrades.

Currently, when upgrades are performed using `oc adm upgrade` CVO will
check if it is a z-level upgrade by comaparing the minor level version
of the desired update with the minor level version of the currently
installed version. The currently installed version is determined by
simply using the first entry in version history however this is
incorrect since that entry will be the version being upgraded to, i.e.
the desired version. A side-effect of this is that when the
`--to-image option` is used, this first entry will have an empty version
string and therefore the z-level upgrade check will always return false
and preconditions are always checked.
Instead, this change iterates through the version history until it finds
the first version with update State of Completed which will be the
currently installed version. If no Completed version is found the
originally installed version is returned.