install/0000_90_cluster-version-operator_02_servicemonitor: Info-level alert for available updates
The purpose of this alert is to give cluster-admins something they can
subscribe to (via Alertmanager notifications) to let them know when
they can initiate an update for a particular cluster.

It is info severity because we want the quick push notification, but
letting your cluster sit on the old version for a few days or a week
or so is unlikely to cause much trouble.  For things like CVE fixes, a
more timely update might be advisable, but the cluster-version
operator currently has no way to make the "critical CVE" vs. "no
fixes, but adds a feature" distinction.  So I'm punting on the
warn-level or so "CVO is getting grumpy about your failure to take the
long-recommended update" alert for now, and just addressing the
info-level "you could update, no pressure".