*: Port from Update to Release for ClusterVersion status
The next/last Actual comparison avoids:

  invalid operation: next.Actual == last.Actual (struct containing []string cannot be compared)

The image pullspec is sufficient for that matching.  The test suite's
expected LastProgress changes are because the shift from:

  {Version: "4.0.1", Image: "image/image:1"}

to:

  {Version: "1.0.0-abc", Image: "image/image:1"}

no longer counts as a LastProgress-bumping change (because we're now
only comparing the Image pullspecs).

Also adjusts the upstream/Cincinnati retrieval and payload loading
logic to preserve the additional metadata that we're now exposing.
Storing a Release on the Operator, instead of extending the previous
releaseVersion/releaseImage pattern, makes it easy to keep track of
payload-loaded metadata, so we can fall-back to upstream metadata for
properties where the payload has no opinion, as described in the
enhancement.

In findUpdateFromConfigVersion, if some crazy upstream service
declared a node with a matching version but no image, we used to say
"sorry, no match" without looking through history.  Now we skip the
useless entry and move on to consider later entries and history.  We
only say "no match" if we have no image pullspec associated with the
requested version available.