lib: Add autogeneration for some resource* functionality
Protecting us from [1,2]:

  converting (v1beta1.Role) to (v1.Role): unknown conversion

Because merging and application are also version-dependent, it's hard
to paper over this in resourceread with automatic type conversion.
Although from [3]:

  Promotes the rbac.authorization.k8s.io/v1beta1 API to v1 with no changes

so all we really need is the apiVersion bump.  Anyhow, with this
commit, I'm doubling down on the approach from 4ee7b07e99 (Add
apiextensions.k8s.io/v1 support for CRDs, 2019-10-22, #259) and
collapsing the readers into two helpers that support all of our types
and return runtime.Object.

From 0a255aba5c (cvo: Use protobuf for sending events and other basic
API commands, 2019-01-18, #90), protobuf is more efficient, so we
should use it where possible.

And because all of this is very tedious to maintain by hand, there's
now a Python generator to spit out all the boilerplate.

[1]: https://github.com/kubernetes/kubernetes/pull/90018
[2]: https://github.com/openshift/cluster-version-operator/pull/420#issuecomment-665960650
[3]: https://github.com/kubernetes/kubernetes/pull/49642