pkg/cvo/status: Raise Operator leveling grace-period to 20 minutes
Reduce false-positives when operators take a while to level (like the
machine-config operator, which has to roll the control plane
machines).  We may want to raise this further in the future, but baby
steps ;).

The previous 10-minute value is from c2ac20fa17 (status: Report the
operators that have not yet deployed, 2019-04-09, #158), which doesn't
make a case for that specific value.  So the bump is unlikely to break
anything unexpected.