pkg/cincinnati/cincinnati: Change "installed" -> "reconciling" for VersionNotFound
When you initiate an update from A -> B, if B is not found in the
current channel, the cluster would start reporting VersionNotFound
after it's first attempt to retrieve recommended updates for B.  That
usually happens before the update completes, and the old "currently
installed version" wasn't a very accurate description of B mid-update.
The new wording relaxes our claim about the cluster state (maybe it
has been reconciled, maybe not) and just talks about the version we're
attempting to reconcile.

The addition of "cluster" is intended to be more explicit about which
version we are discussing [1].  Personally I think the occurence
within the CVO repo's docs and ClusterVersion condition messages is
sufficient context, so I'm agnostic about the additional "cluster".

[1]: https://github.com/openshift/cluster-version-operator/pull/423#discussion_r464350030